style : 환경설정 페이지 UI
