+++ src/features/my/ui/Info.tsx
@@ -1,15 +1,9 @@
 import MyNextArrow from "../../../shared/assets/MyNextArrow.svg";
-import FeedbackIcon from "../../../shared/assets/FeedbackIcon.svg";
 import { useNavigate } from "react-router-dom";
 import { useState } from "react";
 import LogoutModal from "../../../features/auth/ui/LogoutModal";
 
 function Info() {
-  const handleClick = () => {
-    window.location.href =
-      "https://docs.google.com/forms/d/e/1FAIpQLSe-d5MhQrwsRRrk9isYiYVw1afI7a60Xm0IHbxmmAHe8AUiMA/viewform";
-  };
-
   const handleClickUpdate = () => {
     window.location.href =
       "https://youz2me.notion.site/Livith-v-25-04-13-1d402dd0e5fc800dab7fc177f325eade?pvs=4";
@@ -25,19 +19,10 @@ function Info() {
 
   return (
     <div>
-      <img
-        onClick={handleClick}
-        src={FeedbackIcon}
-        className="w-full h-full px-16 py-20 cursor-pointer"
-      />
-
-      <div className="pl-16 pr-18">
-        <div className="mt-14 flex justify-between">
+      <div className="px-16">
+        <div className="mt-30 cursor-pointer ">
           <p className="m-0 text-grayScaleBlack30 text-Body2-md font-medium font-NotoSansKR">
-            버전 정보
-          </p>
-          <p className="m-0 text-grayScaleBlack30 text-Body4-re font-regular font-NotoSansKR">
-            2.0.1
+            알림 설정
           </p>
         </div>
 
@@ -65,8 +50,17 @@ function Info() {
           </button>
         </div>
 
+        <div className="mt-30 flex justify-between">
+          <p className="m-0 text-grayScaleBlack30 text-Body2-md font-medium font-NotoSansKR">
+            버전 정보
+          </p>
+          <p className="m-0 text-grayScaleBlack30 text-Body4-re font-regular font-NotoSansKR">
+            2.1.0
+          </p>
+        </div>
+
         <div
-          className="mt-30 cursor-pointer "
+          className="mt-30 cursor-pointer"
           onClick={() => setIsModalOpen(true)}
         >
           <p className="m-0 text-grayScaleBlack30 text-Body2-md font-medium font-NotoSansKR">
