+++ src/features/my/ui/Nickname.tsx
@@ -7,8 +7,8 @@ function Nickname() {
   const user = useRecoilValue(userState);
   return (
     <>
-      <div className="flex justify-between h-297 inset-0 bg-gradient-to-b from-grayScaleBlack80 to-transparent opacity-100">
-        <div className="mx-16 mt-213 mb-20 w-full flex justify-between items-end items-center">
+      <div className="flex justify-between h-210 inset-0 bg-gradient-to-b from-grayScaleBlack80 to-transparent opacity-100">
+        <div className="mx-16 mt-126 mb-20 w-full flex justify-between items-end items-center">
           <div className="">
             <div className="flex items-center">
               <p className="mr-2 text-grayScaleWhite text-Head1-sm font-semibold font-NotoSansKR">
